+++ CHANGELOG.md
@@ -24,8 +24,9 @@ Expected: September 2026
 ### Features
 
 * New: [NACM External Groups](https://github.com/clicon/clixon/issues/654)
-  Added explicit groupname attribute for cred-mode=NONE, for masquerading of groups
-  See: https://clixon-docs.readthedocs.io/en/latest/netconf.html#external-groups
+  * Added explicit groupname attribute for cred-mode=NONE, for masquerading of groups
+    * Added `-g <group>` option to the CLI
+  * See: https://clixon-docs.readthedocs.io/en/latest/netconf.html#external-groups
 * Reentrant (thread-safe) parsers: XML, YANG, XPATH, JSON, etc.
 
 ### API changes